initial pass of dynamic geofence support complete

+++ Trackit/StaticRadiusGeofence+CoreDataProperties.swift
@@ -2,9 +2,8 @@
 //  StaticRadiusGeofence+CoreDataProperties.swift
 //  Trackit
 //
-//  Created by Richard Broberg on 12/23/16.
+//  Created by Richard Broberg on 12/24/16.
 //  Copyright © 2016 Brobasino. All rights reserved.
-//  This file was automatically generated and should not be edited.
 //
 
 import Foundation
@@ -19,6 +18,6 @@ extension StaticRadiusGeofence {
 
     @NSManaged public var latitude: Double
     @NSManaged public var longitude: Double
-    @NSManaged public var radius: Double
+    @NSManaged public var radius: Float
 
 }
